feat: moved types from chat components to index file in types
